# core/docs/swagger.yaml
basePath: /api/v1
...
host: localhost:8080
info:
  contact: {}
  description: High-performance BFF for Omsu university schedule.
  title: Omsu Schedule Mirror API
  version: "1.0"
paths:
...
  /schedule/auditory/{id}/ical:
    get:
      description: Returns the schedule for an auditory in .ics format. If ICAL_ACCESS_TOKEN
        is configured, pass it as the token query parameter.
      parameters:
      - description: Auditory ID
        in: path
        name: id
        required: true
        type: integer
      - description: Optional iCal access token
        in: query
        name: token
        type: string
      produces:
      - text/calendar
      responses:
        "200":
          description: iCalendar payload
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Export auditory schedule as iCalendar
      tags:
      - Schedules
...
  /schedule/group/{id}/ical:
    get:
      description: Returns the schedule for a group in .ics format. If ICAL_ACCESS_TOKEN
        is configured, pass it as the token query parameter.
      parameters:
      - description: Group ID (real_group_id)
        in: path
        name: id
        required: true
        type: integer
      - description: Optional iCal access token
        in: query
        name: token
        type: string
      produces:
      - text/calendar
      responses:
        "200":
          description: iCalendar payload
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Export group schedule as iCalendar
      tags:
      - Schedules
...
  /schedule/tutor/{id}/ical:
    get:
      description: Returns the schedule for a tutor in .ics format. If ICAL_ACCESS_TOKEN
        is configured, pass it as the token query parameter.
      parameters:
      - description: Tutor ID
        in: path
        name: id
        required: true
        type: integer
      - description: Optional iCal access token
        in: query
        name: token
        type: string
      produces:
      - text/calendar
      responses:
        "200":
          description: iCalendar payload
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Export tutor schedule as iCalendar
      tags:
      - Schedules
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-Admin-Key
    type: apiKey
swagger: "2.0"